chore: Update image sizes and optimize performance
The commit updates the image sizes in the README file and optimizes the performance of the application by reducing the maximumWarning and maximumError values for the initial budget in the angular.json file. It also adds conditional rendering for the canvas elements in the dashboard.component.html file based on the isBrowser flag. Additionally, it initializes the isBrowser flag in the ngOnInit method of the dashboard.component.ts file. Finally, it updates the environment files to set the production flag accordingly and enables production mode in the main.ts file when the environment is set to production.